+++ kogito-drools/drools-core/src/main/java/org/drools/impl/StatefulKnowledgeSessionImpl.java
@@ -129,7 +129,7 @@ public class StatefulKnowledgeSessionImpl
         return this.session.getEntryPointNode();
     }    
 
-    public int getId() {
+    public long getId() {
         return this.session.getId();
     }
 
@@ -844,7 +844,7 @@ public class StatefulKnowledgeSessionImpl
 		this.session.queueWorkingMemoryAction(action);
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.session.setId(id);
 	}
